docker_finn/requirements.txt
confluent-kafka==2.1.1
websocket-client==1.5.2